--- a/.gitignore
+++ b/.gitignore
@@ -7,3 +7,4 @@ build
 .idea
 .iml
 local.properties
+.gradle