--- a/iNaturalist/src/main/res/layout/observation_list_item.xml
+++ b/iNaturalist/src/main/res/layout/observation_list_item.xml
@@ -6,8 +6,7 @@
     android:paddingLeft="15dp"
     android:paddingRight="15dp"
     android:layout_width="fill_parent"
-    android:layout_height="70dp">
-
+    android:layout_height="wrap_content">
 
     <ImageView
         android:id="@+id/checkbox"
@@ -136,30 +135,28 @@
             style="@android:style/TextAppearance.Medium"
             android:text="Species Guess" />
 
-
-        <LinearLayout
-            android:orientation="horizontal"
-            android:gravity="center_vertical"
+        <ImageView
             android:layout_below="@id/species_guess"
             android:layout_alignParentLeft="true"
+
+            android:id="@+id/location_icon"
+            android:layout_width="16dp"
+            android:layout_height="wrap_content"
+            android:src="@drawable/ic_location_on_black_24dp"
+            android:tint="#6F6F6F"
+            />
+
+        <TextView
+            android:layout_below="@id/species_guess"
+            android:layout_toRightOf="@id/location_icon"
+            android:id="@+id/place_guess"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content">
-            <ImageView
-                android:id="@+id/location_icon"
-                android:layout_width="16dp"
-                android:layout_height="wrap_content"
-                android:src="@drawable/ic_location_on_black_24dp"
-                android:tint="#6F6F6F"
-                />
-            <TextView
-                android:id="@+id/place_guess"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:gravity="center_vertical"
-                android:maxLines="1"
-                android:ellipsize="end"
-                android:text="San Francisco" />
-        </LinearLayout>
+            android:layout_height="wrap_content"
+            android:maxLines="1"
+            android:ellipsize="end"
+            android:gravity="center_vertical"
+            android:layout_alignBottom="@id/location_icon"
+            android:text="San Francisco" />
 
     </RelativeLayout>