… iNaturalist/src/main/res/values/strings.xml