I noticed the observation list has some issues if you enable the system-level accessibility features of increasing font size globally or increasing "display" (e.g. everything) size globally. 

Two examples: left image is a simple "increase fonts globally" screen cap  (Settings>Display>Font Size), and right is a "scale all display elements" screen cap (Settings>Display>Display Size) (this setting is IMO more common for poor eyesight). I've marked up the right capture to show the issues. I'm ignoring the tabview issues and just focusing on the list view for now. 

<img width="200" alt="Screen Shot 2020-05-19 at 11 11 33 PM" src="https://user-images.githubusercontent.com/305380/82410063-17ba7680-9a3d-11ea-9459-319a7a3f1d08.png"><img width="200" alt="Screen Shot 2020-05-20 at 1 57 20 AM" src="https://user-images.githubusercontent.com/305380/82413443-57845c80-9a43-11ea-8bf5-19256c95f84f.png">


After I looked into the code, I realized there was also a performance issue. It's nesting Linear and Relative layouts 3-4 layers deep, causing lots of layout work for every single list item. Lots of this nesting was for centering and alignment, so we can just switch the layout over to a Constraint layout and get two benefits - more functionality and more performant layout. 

The commits here prove the concept out and I can finalize if this looks OK to maintainers. Performance-wise there are still improvements that could be made, but this looks much better already: 

| Metric           | Original Nested Layouts          | New ConstraintLayout          |
|---------------------------------|--------------|--------------|
| Total frames rendered           | 199          | 173          |
| Janky frames                    |  32 (16.08%) |  22 (12.72%) |
| 50th percentile                 |  11ms        |  10ms        |
| 90th percentile                 |  24ms        |  20ms        |
| 95th percentile                 |  26ms        |  24ms        |
| 99th percentile                 |  300ms       |  105ms       |
| Number Missed Vsync             | 5            | 2            |
| Number High input latency       | 0            | 0            |
| Number Slow UI thread           | 16           | 8            |
| Number Slow bitmap uploads      | 2            | 1            |
| Number Slow issue draw commands | 24           | 16           |



Some example screenshots varying system font/size settings wildly. Notice that we have consistent image/text alignment, no vertical cropping, and the horizontal text for species guess and location guess is now being automatically trimmed if it begins to overlap the right-most content. 

<img width="200" alt="Screen Shot 2020-05-20 at 2 14 44 AM" src="https://user-images.githubusercontent.com/305380/82411682-3a9a5a00-9a40-11ea-97cc-a188c92f9b75.png"><img width="200" alt="Screen Shot 2020-05-20 at 2 15 53 AM" src="https://user-images.githubusercontent.com/305380/82411690-3d954a80-9a40-11ea-8f5b-5119ca417819.png"><img width="200" alt="Screen Shot 2020-05-20 at 2 16 13 AM" src="https://user-images.githubusercontent.com/305380/82411697-40903b00-9a40-11ea-8122-6490893cfe82.png">