@kueda 2020-05-20
FWIW, from a product perspective I think this is great. It would be better to make systemic rather than incremental improvements in both accessibility and performance, but I guess we'll take what we can get.

Technically, I defer to @budowski, so I'm assigning this to you, Yaron.