@hamiltont 2020-05-20
Thanks @kueda 👍 Most of the app behaves well with the system font/display scaling, so this PR is just about fixing an outstanding issue. It's not my intention to start an (unneeded) app-wide UI redesign

Main reason I'm asking for feedback is to know if the technical team is OK with me changing up the list item layout for such an integral part of the application. I think it's an improvement in all dimensions (performance, featureset, accessibility), but want to confirm agreement before I put in work to finalize the PR.