@hamiltont 2020-05-21
Ok, this should be good to review. I tried to leave the Adapter alone as much as possible and focus effort entirely on the UI. It should be identical to prior layout approach, but now with these improvements: 

1. No more text cropping or overlaps
2. Proper accessibility meta-text
3. No linter warnings
4. Right-to-left support
5. Slightly faster. Note I had to add one allocation inside `getView` to avoid rewriting logic in the adapter. systrace reports it's still faster, but that one alloc does slow us down a bit. From what I can tell it's possible to optimize adapter code much, so will leave removing this as a TODO 

Of interest, cropping got a bit better with these new commits. In the edge case where the progress indicator shows, the species guess now properly crops to either the date or progress indicator, whichever comes first in the layout. Also, I added margins to the cropping, so any cropped text is not squished right next to the view we are trying to prevent it from overlapping. 

<img width="452" alt="Screen Shot 2020-05-20 at 11 27 12 PM" src="https://user-images.githubusercontent.com/305380/82524800-d8079380-9afd-11ea-9798-0da2e9510077.png">

Additionally, RTL works as expected and I've confirmed the UI fixes appear to function as expected (aligns, crops, etc) even with funny scaling of font/display: 

![Screenshot_1590034951](https://user-images.githubusercontent.com/305380/82524933-274dc400-9afe-11ea-8ac4-0ad0092ac789.png)

Unmarking as draft, IMO this is ready. Could tweak more (tab views and titlebar) but don't want this to overgrow